Moved Task class to Class
